requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf2
google-generativeai
